refactor: pull Scenario code out of main module
