+++ tests/end_to_end_cases/management_api.rs
@@ -7,7 +7,7 @@ use test_helpers::assert_contains;
 
 use crate::common::server_fixture::ServerFixture;
 
-use super::util::{
+use super::scenario::{
     create_readable_database, create_two_partition_database, create_unreadable_database, rand_name,
 };
