+++ tests/end_to_end_cases/management_cli.rs
@@ -4,7 +4,7 @@ use test_helpers::make_temp_file;
 
 use crate::common::server_fixture::ServerFixture;
 
-use super::util::{create_readable_database, rand_name};
+use super::scenario::{create_readable_database, rand_name};
 
 #[tokio::test]
 async fn test_writer_id() {
