+++ tests/end_to_end_cases/mod.rs
@@ -3,7 +3,7 @@ pub mod management_api;
 pub mod management_cli;
 pub mod read_api;
 pub mod read_cli;
+pub mod scenario;
 pub mod storage_api;
-pub mod util;
 pub mod write_api;
 pub mod write_cli;
